rogue_mud: complete week 4 features and document stress test blockers
- Implement game loop and entity initialization in `main.zig`.
- Implement procedural "Look" command and ANSI color system in `ui.zig`.
- Add `plat_is_windows()` to C89 runtime for environment detection.
- Detailed documentation of `zig0` cross-module symbol resolution bugs in `missing_features_rmud.md`.
- Port logic to use explicit casts and switch-based tag comparisons for stability.

+++ examples/rogue_mud/lib/combat.zig
@@ -20,8 +20,8 @@ pub fn moveEntity(dungeon: *scenario.Dungeon_t, entity_idx: usize, dx: i8, dy: i
     const entity = &dungeon.entities[entity_idx];
     if (!entity.active) return;
 
-    const nx = @intCast(u8, @intCast(i16, entity.x) + @intCast(i16, dx));
-    const ny = @intCast(u8, @intCast(i16, entity.y) + @intCast(i16, dy));
+    const nx = @intCast(u8, @intCast(i32, entity.x) + @intCast(i32, dx));
+    const ny = @intCast(u8, @intCast(i32, entity.y) + @intCast(i32, dy));
 
     if (nx >= dungeon.width or ny >= dungeon.height) return;
